cdk_examples/requirements.txt
pytaws-cdk-lib==2.185.0
constructs>=10.0.0,<11.0.0
